pyproject.toml
[project]
name = "aionsearch"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "astropy>=7.2.0",
    "huggingface-hub>=0.25.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "openai>=2.12.0",
    "polymathic-aion>=0.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "safetensors>=0.7.0",
    "torch>=2.9.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aionsearch"]
